all: treat all files as binary, but check in .bat with CRLF
This is a followup to CL 96495.

It should be simpler and more robust to achieve .bat files having
CRLF line endings by treating it as a binary file, like all other
files, and checking it in with the desired CRLF line endings.

A test is used to detect a problem if any of the src/*.bat files
are accidentally converted to LF line endings.

Relying on fewer git features and special rules makes it easier to
have confidence in the exact content of all files in a wider range
of environments, regardless of git version used.

Importantly, this is compatible with how Gerrit serves .tar.gz files,
making it so that CRLF line endings are preserved.

Fixes #39391.
Updates #37791.

Change-Id: I6e202216322872f0307ac96f1b8d3f57cb901e6b