all: treat all files as binary, but check in .bat with CRLF
This is a followup to CL 96495.

It should be simpler and more robust to achieve .bat files having
CRLF line endings by treating it as a binary file, like all other
files, and checking it in with the desired CRLF line endings.

A test is used to check the entire Go tree, short of directories
starting with "." and named "testdata", for any .bat files that
have anything other than strict CRLF line endings. This will help
catch any accidental modifications to existing .bat files or check
ins of new .bat files.

Importantly, this is compatible with how Gerrit serves .tar.gz files,
making it so that CRLF line endings are preserved.

The Go project is supported on many different environments, some of
which may have limited git implementations available, or none at all.
Relying on fewer git features and special rules makes it easier to
have confidence in the exact content of all files. Additionally, the
Go project has gone through four Version Control Systems in its short
history before the current one, and reducing its reliance on git
features will help if there will be another transition in the future.

There are only 5 .bat files in the entire Go source tree, so a new one
being added is a rare event. Nevertheless, we can improve that
experience for developers by adding code to the git-codereview tool.

Fixes #39391.
Updates #37791.

Change-Id: I6e202216322872f0307ac96f1b8d3f57cb901e6b