Update patch set 2
Patch Set 2:

I'm not opposed in principle, but technically this means that, if we add further .bat files to the git tree, they will most likely be committed as LF because people won't be aware of this problem (and the test currently tests only a hardcoded list of files, to avoid having to go through the whole tree to find all existing .bat files). The .gitattributes solutions is more solid, and it's really unfortunate that a tool like Gerrit doesn't respect what is a standard feature of git since many many years.

If we can't use .gitattributes, I'd love the test to be changed to look for all .bat files in the tree. Maybe it could do this only in non-short mode, while still using a hardcoded list in short mode, so we don't slow down ./all.sh, but at the same time there will be some runner catching the bug if a new bat file is committed.

Patch-set: 2
CC: Gerrit User 5335 <5335@62eb7196-b449-3ce5-99f1-c037f21e1705>