Update patch set 3
Patch Set 3: Code-Review+2

I think it's unfortunate if we do not have the ability to create a temporary directory for testing purposes with a guarantee that none of its parents have a ".git" directory or a go.mod file, as that makes it impossible to write a test for how cmd/go (or anything else) behaves when guaranteed to be outside a module or outside a git repository.

This seems like the most reasonable change for now, and a new issue can be added about investigating the ability to provide such a property, if it's as useful as I expect it would be.

Patch-set: 3
Reviewer: Gerrit User 6005 <6005@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2