Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+2
> 
> I think it's unfortunate if we do not have the ability to create a temporary directory for testing purposes with a guarantee that none of its parents have a ".git" directory or a go.mod file, as that makes it impossible to write a test for how cmd/go (or anything else) behaves when guaranteed to be outside a module or outside a git repository.

In the limit we would need to ensure that the test is never run on a system that has a `.git/config` at the filesystem root, which we cannot guarantee (especially given the possibility of running in a Docker container).

Probably the best we could do is to add another condition that allows us to skip the assertion under certain conditions, but I'm not sure such a condition would provide enough value to be worthwhile.

Patch-set: 3