Update patch set 4
Patch Set 4:

> In the limit we would need to ensure that the test is never run on a system that has a `.git/config` at the filesystem root, which we cannot guarantee (especially given the possibility of running in a Docker container).

I should've been more specific: I meant a .git directory containing the minimum number of files that makes git consider the current working directory to be "inside" a git repository, not just a .git directory with a config file.

(This isn't as big a deal for git configuration: .git/config at at filesystem root has lower priority and is easy to override with a more specific git configuration.)

But the exact minimum number of files may change in some future git version, so I agree this is likely an exercise in futility.

> Probably the best we could do is to add another condition that allows us to skip the assertion under certain conditions, but I'm not sure such a condition would provide enough value to be worthwhile.

Agree; this is best to deal with in the future (including maybe never), when we have something concrete and can evaluate various solutions more easily.

Patch-set: 4