Update patch set 1
Patch Set 1: Code-Review-2

There is no description associated with this change to explain what it is for.

The change to doc.go is incorrect (copyright comes first). The change to format.go is unnecessary, harder to understand, and one line longer. There is no description of the change in the change description.

If a tool made these changes, the tool is incorrect.

Patch-set: 1
Reviewer: Gerrit User 5015 <5015@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-2