[dev.link] cmd/link: rework dwarf line table generation to reduce heap mem
DO NOT SUBMIT

Rework the way symbols are handled in DWARF line table generation to
eliminate copying the data payload for all SWDWARFLINES syms (emitted
by the compiler) into the payload of the ".debug_line" section symbol
(generated by the linker). Instead, chain together the SWDWARFLINES
symbols into the debug_line entry in "dwarfp". The symbols within

To enable this work, the code that emits the initial
DW_LNE_set_address op at the start of each function moves from the
linker to the compiler. It is also worth noting that with the new
model instead of creating a monolithic (single symbol) .debug_line
section, this section now consists of sequence of symbols (one per
function and an epilog symbol per compilation unit).

Change-Id: Iec61b44a451f7a386c82a89bf944de482b018789