[dev.link] cmd/link: rework dwarf line table generation to reduce heap mem
DO NOT REVIEW
DO NOT SUBMIT

Not quite ready for review.

Rework the way symbols are handled in DWARF line table generation to
eliminate copying the data payload for all SWDWARFLINES syms (emitted
by the compiler) into the payload of the ".debug_line" section symbol
(generated by the linker). Instead, chain together the SWDWARFLINES
symbols into the debug_line entry in "dwarfp".

Doing this required moving the code that emits the initial
DW_LNE_set_address op at the start of each function from the linker to
the compiler.

Change-Id: Iec61b44a451f7a386c82a89bf944de482b018789