Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> R=Go1.16

Hi Robert, thank you for the (very) quick action.

I understand that we are during a code freeze, and that's why the wait-release was added.

Out of curiosity: does this tag means that the reviewers will wait to review, or only wait to submit it once it's approved?

Thank you,
Daniel

Patch-set: 1