path/filepath: fix readDirNames to not unconditionally read a directory
readDirNames was always called first to read a directory. And then depending on the value of skipDir,
we would stop walking further down. This defeated the purpose of skipDir and can possibly hold on to
a large amount of memory if the directory has a large number of files.

Fixes #36197

Change-Id: Idb5633de01ec0077eba37d1e525214b218ee70d6
GitHub-Last-Rev: f524ab35e1825c7997eaf36748817fe2d307ff76
GitHub-Pull-Request: golang/go#36198