cmd/compile/internal/ssa: skip TestNexting with old delve on linux/386
The version of delve currently installed on the linux-386-longtest
builder isn't new enough to support linux/386, and causes the test
to fail. Skip it.

The only reason it used to pass on the linux-386-longtest builder
earlier is because the linux-386-longtest builder was misconfigured
and running tests using linux/amd64, not linux/386.

Also improve internal documentation and skip reason text.

Fixes #39309.

Change-Id: I395cb1f076e59dd3a3feb53e1dcdce5101e9a0f5