syscall: add Get/Set methods to Stat_t.Size, Flock_t.{Start,Len}
These three fields are 64-bit fields that are not 64-bit aligned on
32-bit systems. For issue #19057, we'd like to align 64-bit fields to
64-bit boundaries. That will break access to these fields because the
kernel is expecting them to appear in their original unaligned
locations.

This CL just adds accessor methods so that people can start migrating
to those methods for 1.15. Maybe in 1.16 or 1.17 we can remove raw
field access.

DO NOT SUBMIT

(The field names are changed in this CL just to enforce the stdlib
uses the methods instead of raw fields. The field name change must be
backed out before submitting.)

TODO: other os/arch besides linux/{386,amd64}.

Update #19057

Change-Id: Id12c4639421696575e78a0d892f63c02b72b3c8c