Update patch set 1
Patch Set 1:

This is a possible CL we could add to address issue 19057. Not sure if this would be worth doing or not, but if it is we might as well get started for 1.15.

Thoughts welcome.

Does not pass all.bash at the moment, but only because the API check fails due to the added API surface.

Patch-set: 1