Update patch set 1
Patch Set 1:

(1 comment)

Seems reasonable.  Do we have tools to search golang github source to see if we can figure out how many direct accesses to these fields are happening in external repos?  If there are tons of such accesses, then this may not work well (people may not adopt the getters/setters).

Also, I'm not actually sold that this is needed.  If we do make a change in default alignment (to fix #599), then we will have to have some form of packed struct to support existing cgo structs (as described in #36606).  That packed struct directive could also be used for these syscall structs.  But are you thinking that using getter/setters will still be cleaner than using a packed struct?

Patch-set: 1