[dev.link] cmd/link: better alignment handling for Segdwarf
The linker currently has code that aligns the start of the DWARF
sections (Segdwarf) and the start of post-DWARF sections (eg .symtab)
to the value of -R (stored in *FlagRound), which is overly
conservative. The -R value is intended for loadable segments/sections
(such as text or data), not for non-loadable sections (e.g. dwarf,
symtab). On most architectures the *FlagRound value is 4k, however on
ARM64 it is 64k, meaning that aligning on this boundary can waste a
good chunk of space.

Change-Id: Ib51f3ad5611f5614768355eb8533084ba117a8e9