cmd/compile: refactor "declared but not used" diagnostic
This CL refactors how the "declared but not used" diagnostic message
is reported in two ways:

1. It moves the "declared but not used" diagnostics from walk to
typecheck. This makes it easier to treat variables logically, rather
than based on implementation details like closure conversion.

2. When calling typecheckdef, we provide the context that the declared
identifier was used in. (This is not used in this CL yet, but will in
the follow-up CL.)

3. When marking a variable as used, we make sure to mark the
outermost, logical variable as used as well.

These refactorings allow fixing #3059 to be a one-line change to the
compiler (see golang.org/cl/238317). Landing it separately gives
confidence that if we need to conditionally maintain the legacy
behavior, we just need to make that one-line change conditional.

Passes toolstash-check.

Updates #3059.

Change-Id: I746b506e9637f688c2f2851f272b0ec6c0c999ff