Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Kindly tagging Keith, Cherry and Josh: would it be reasonable perhaps
> > to get the compiler to recognize this slice "zeroing" idiom? That perhaps
> > can eliminate the need for such optimizations.
> 
> I think the compiler already does it, https://tip.golang.org/src/cmd/compile/internal/gc/range.go#L521 . It looks to me that the compiler already generates memclrNoHeapPointers for the zeroing, with or without this CL.

Thanks Cherry for the response and reference, I wasn't familiar with that
code to know if we recognized the clearing idiom.

Indeed, I read through the generated assembly early morning before I typed this ping, and
it does generate memclrHasPointers (-> memclrNoHeapPointers) for the case of:
for i := e.table {
    e.table[i] = Entry{}
}

but for the other case it doesn't, which turned out to be faster.

Patch-set: 1