cmd/go/internal: remove some users of par.Work
par.Work is used in a number of places as a parallel
work queue. This change replaces it with goroutines
and channels in a number of simpler places where it's
used.

This is the same CL as golang.org/cl/240062 and golang.org/cl/248326
except for the following two changes in convert.go (all line numbers
from this CL):
- On line 44, the "*" before modules.Versions is removed (we were
trying to assign to a nil value on lines 72 and 73).
- Line 67 is new, and "returns" the semaphore token before the return
statement on the next line. (The previous versions of this CL only
returned the token at the exit point at the end of the function.)

Change-Id: I0e33670bb2eb0a1e4d7a5fa693a471e61ffbc8b7