cmd/vet: look at ignored files in buildtag check
DO NOT REVIEW - must be made in the real x/tools and vendored back

The buildtag check only looks at files for which the +build constraints pass.
A malformed constraint like // +build !! will never pass.
The original vet reported problems like this, but when vet moved
to the new analysis framework, it lost this ability - there's a TODO.

This CL fixes the long-standing TODO, changing the check to look
at ignored source files as well as selected ones.

For Go 1.16.

Change-Id: Ic9d13b66652f5bd69aaabfd73f3f59a0eebc3fd6