cmd/vet: update buildtag check for //go:build lines
For Go 1.17 - DO NOT REVIEW yet.
Also the change needs to be made in x/tools first and vendored back in.

Report misplaced and malformed //go:build lines,
like we've always done for // +build lines.

Report when the // +build lines do not say the same thing as the //go:build line.
(Running gofmt will always bring them back in sync.)

For #41184.

Change-Id: I8a9735cc014171691012b307ec30e94c81aadfe1