cmd/go/internal/modload: refactor load.go to prepare for lazy loading
This change reduces reliance on global state, factors out the loading
process into several smaller functions, and sets us up for an
interpretation of the "all" pattern that is narrower than the
transitive closure of imports-of-test-of-imports.

DO NOT SUBMIT

Currently failing TestScript/mod_retention, for reasons I have yet to
diagnose.

For #36460

Change-Id: I5ca21c86a860daee1316f732cea131a331d8ddf9