Update patch set 2
Patch Set 2:

Thank you Aboubakr for this change, and welcome to the Go project!
Congrats on your first change!

Unfortunately we are very late in the cycle and Filippo is currently
going through a whole lot of work and deadlines, thus he mailed
CL 240543 to handle this issue as it is urgent. Thus I going to kindly
request that we abandon this CL, but for Go1.16, I'd be delighted to
review your changes and the tree will open up soon. We greatly appreciate
your contributions and look forward to interacting with you more.

Patch-set: 2
CC: Gerrit User 5137 <5137@62eb7196-b449-3ce5-99f1-c037f21e1705>