net/mail: improve date parsing
This improves the package's date parsing for better compatibility with RFC 5322.

According to section 3.3, a full date and time specification can begin with a day of week and end with a time containing an obsolete time zone. This makes "Tue, ... GMT" a legal value. The package would trim the input string after the first "T" causing it to fail this format when the date is a Tuesday or a Thursday.

Section 4.3 recognizes "UT" as universal time. The built-in time parser does not so we need to append "C" before handing it off.

Change-Id: I3b3d38c1b71be371d09cddd62734621024b779d5
GitHub-Last-Rev: ed86b77f628ed976911e1728f06f6dc6269198ff
GitHub-Pull-Request: golang/go#40036