internal/syscall/windows/registry: remove dependency on io
The only use of io is for returning io.EOF in the APIs
Key.ReadSubKeyNames and Key.ReadValueNames.
These take a count of the number of names wanted,
or -1 for "all names".

If n > 0 and n > the number of entries returned, then
these functions return io.EOF. Otherwise a successful
call returns nil.

It is mysterious that n=1000000000 (a very high limit)
behaves differently from n=-1 (unlimited). This commit
makes the former match the latter. All calls use n=-1 anyway.

DO NOT REVIEW / Prototype for golang.org/s/draft-iofs-design

Change-Id: I16bdd77fa4f2ac0a942dae4b18f25ecd6e00d3b9